Use the latest release of RancherOS during a Packer Build.
Added a python-based docker-entrypoint that uses the GitHub API to determine the latest
released version of RancherOS.  The entrypoint then retrieves the checksum from the project and
the ISO url.  This two parameters are added to the Packer command line arguments.

Removed Fedora and CentOS based images as they are no longer needed.

Modified the Dockerfile for the new docker-entrypoint and limited the files copied to the container.