Changes to support embedded container image
Created an on and offline build.  Modified `rancheros.json` accordingly to support this.
Including the proper commands to _save_ the image locally after the installation of RancherOS

Modified `docker-compose.yml` to support the building of both on and offline.
Needed to modify the VNC ports since two builders were running at the same time and being blocked
with no available port.

Increased the amount of memory required to 4GB in offline mode.  This number could be lowered just what it was
tested with.  Otherwise there is a race condition that the compose operation happens before the preload can complete.